+++ src/features/SupplyChainFlow/Assets/AssetReducers.js
@@ -62,7 +62,7 @@ const AssetReducers = (state = INITIAL_STATE, action) => {
             return Object.assign({}, state, {
                 ...state,
                 assetFetched: true,
-                showPasswordModal: true,
+                // showPasswordModal: true, commented out to fascilitate debugging
                 selectedAsset: action.selectAsset
             })
